refactor: 불필요한 컨트롤러 import 제거, 파일명 수정, reportCount min값 0 지정, isDanger 불필요한 쓰기 작업 유도 제거
