+++ src/app.module.ts
@@ -3,7 +3,6 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { AuthModule } from './auth/auth.module';
 import { PrismaModule } from './prisma/prisma.module';
-import { ReportController } from './report/report.controller';
 import { ReportService } from './report/report.service';
 import { ReportModule } from './report/report.module';
 import { JwtGuard } from './auth/jwt.guard';
@@ -11,7 +10,7 @@ import { APP_GUARD } from '@nestjs/core';
 
 @Module({
   imports: [AuthModule, PrismaModule, ReportModule],
-  controllers: [AppController, ReportController],
+  controllers: [AppController],
   providers: [
     AppService,
     ReportService,
